v3.0.10

+++ browserpicker/domain/src/main/java/browserpicker/domain/usecase/history/GetHistoryOverviewUseCase.kt
@@ -15,7 +15,7 @@ import javax.inject.Inject
 
 @Keep
 data class HistoryOverview(
-    val totalCount: Int,
+    val totalCount: Long,
     val groupCounts: List<DomainGroupCount>,
     val dateCounts: List<DomainDateCount>,
 )
